chore(docs): moved di example directive to separate file
